transient, namespace, liveload, resources, cmd, gpt etc

+++ share/xmonad/appmenu.sh
@@ -1,4 +1,11 @@
 #!/bin/sh
 
-killall jgmenu
-jgmenu --at-pointer
+
+if [ -x /usr/bin/xfce4-appfinder ]; then
+    xfce4-appfinder
+elif [ -x /usr/bin/jgmenu ]; then
+    killall jgmenu
+    jgmenu --at-pointer
+else
+    dmenu_run
+fi
